--- src/test/java/ru/yandex/practicum/filmorate/controller/filmTests/FilmrControllerTest3.java
@@ -1,41 +1,0 @@
-package ru.yandex.practicum.filmorate.controller.filmTests;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import ru.yandex.practicum.filmorate.controller.FilmController;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.model.Film;
-
-import java.time.LocalDate;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@SpringBootTest
-class FilmrControllerTest3 {
-    @Autowired
-    private FilmController filmController;
-    Film film;
-
-    @Test
-    public void shouldThrowExceptionWhenUpdateUnknownFilm() {
-        film = Film.builder()
-                .name("nisi eiusmod")
-                .description("adipisicing")
-                .releaseDate(LocalDate.of(1967, 3, 25))
-                .duration(100)
-                .build();
-
-        film.setId(1000);
-
-        NotFoundException e = assertThrows(
-                NotFoundException.class, () -> filmController.updateFilm(film));
-
-        assertEquals("404 NOT_FOUND \"Film with name = \"nisi eiusmod\" don't find\"", e.getMessage());
-        assertEquals(1,
-                filmController.getAllFIlms().size(),
-                "Size Equal Test");
-    }
-
-}
