--- src/test/java/ru/yandex/practicum/filmorate/controller/userTests/UserControllerTest3.java
@@ -1,40 +1,0 @@
-package ru.yandex.practicum.filmorate.controller.userTests;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.context.SpringBootTest;
-import ru.yandex.practicum.filmorate.controller.UserController;
-import ru.yandex.practicum.filmorate.exception.NotFoundException;
-import ru.yandex.practicum.filmorate.model.User;
-
-import java.time.LocalDate;
-
-import static org.junit.jupiter.api.Assertions.assertEquals;
-import static org.junit.jupiter.api.Assertions.assertThrows;
-
-@SpringBootTest
-class UserControllerTest3 {
-    @Autowired
-    private UserController userController;
-    User user;
-
-    @Test
-    public void shouldThrowExceptionWhenUpdateUnknownUser() {
-        user = User.builder()
-                .login("dolore")
-                .name("Nick Name")
-                .email("mail@mail.ru")
-                .birthday(LocalDate.of(1946, 8, 20))
-                .build();
-
-        user.setId(1000);
-
-        NotFoundException e = assertThrows(
-                NotFoundException.class, () -> userController.updateUser(user));
-
-        assertEquals("404 NOT_FOUND \"User don't find\"", e.getMessage());
-        assertEquals(3,
-                userController.getAllUser().size(),
-                "Size Equal Test");
-    }
-}
